line chart: drawable and data conversion
This change introduces a concept of a drawable; a special layout that
can render content such as lines and text.

There are two types of drawable; just drawable and data drawable. One
can render its content without actual data but can only render with
metadata and another can render its content with actual data. An example
of the former would be something like axis which only has to know
viewBox and domSize in order to render content and one for latter would
be actual line chart.

For the DataDrawable, the library attempts to minimally spend CPU cycles
on data coordinate system conversion.